# Make constants in PackedVector constexpr